requirements.txt
python-telegram-bot == 20.2
aiohttp == 3.8.4